Update main.py for new python 3.12.4

+++ main.py
@@ -68,14 +68,14 @@ def mouse_bez(init_pos, fin_pos, deviation, speed):
 
     # bezier centre control points between (deviation / 2) and (deviaion) of travel distance, plus or minus at random
     control_1 = (
-    init_pos[0] + choice((-1, 1)) * abs(ceil(fin_pos[0]) - ceil(init_pos[0])) * 0.01 * randint(deviation / 2,
+    init_pos[0] + choice((-1, 1)) * abs(ceil(fin_pos[0]) - ceil(init_pos[0])) * 0.01 * randint(int(deviation / 2),
                                                                                                deviation),
-    init_pos[1] + choice((-1, 1)) * abs(ceil(fin_pos[1]) - ceil(init_pos[1])) * 0.01 * randint(deviation / 2, deviation)
+    init_pos[1] + choice((-1, 1)) * abs(ceil(fin_pos[1]) - ceil(init_pos[1])) * 0.01 * randint(int(deviation / 2), deviation)
     )
     control_2 = (
-    init_pos[0] + choice((-1, 1)) * abs(ceil(fin_pos[0]) - ceil(init_pos[0])) * 0.01 * randint(deviation / 2,
+    init_pos[0] + choice((-1, 1)) * abs(ceil(fin_pos[0]) - ceil(init_pos[0])) * 0.01 * randint(int(deviation / 2),
                                                                                                deviation),
-    init_pos[1] + choice((-1, 1)) * abs(ceil(fin_pos[1]) - ceil(init_pos[1])) * 0.01 * randint(deviation / 2, deviation)
+    init_pos[1] + choice((-1, 1)) * abs(ceil(fin_pos[1]) - ceil(init_pos[1])) * 0.01 * randint(int(deviation / 2), deviation)
     )
 
     xys = [init_pos, control_1, control_2, fin_pos]
